Allows For Increased Control Over LED Strips
Code allows each of the 8 LED strips to be controlled individually, in terms of color using RGB values and on and off in general. The variables a, b, c, d, e, f, g, h, redValue, greenValue, abd blue Value are meant to be set by the user to determine the color and on and off state of each LED strip.